fix hm02-express
